requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0

# Swiss Ephemeris for planetary calculations
pyswisseph==2.10.3.2

# Database
supabase==2.0.0
psycopg2-binary==2.9.9

# API and HTTP
requests==2.31.0
python-dotenv==1.0.0

# OpenAI for automated event collection
openai==1.6.1

# Email
# Note: smtplib and email are built-in Python modules

# Date/time handling
python-dateutil==2.8.2

# Type checking (optional but recommended)
mypy==1.7.1
types-requests==2.31.0.10
